+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Accept two digit country codes
+
 ## [1.22.0] - 2024-06-03
 
 ### Fixed